+++ server/src/main/java/org/opensearch/index/translog/InternalTranslogManager.java
@@ -58,7 +58,8 @@ public class InternalTranslogManager implements TranslogManager {
         TranslogEventListener translogEventListener,
         LifecycleAware engineLifeCycleAware,
         TranslogFactory translogFactory,
-        BooleanSupplier startedPrimarySupplier
+        BooleanSupplier startedPrimarySupplier,
+        TranslogOperationHelper translogOperationHelper
     ) throws IOException {
         this.shardId = shardId;
         this.readLock = readLock;
@@ -71,7 +72,7 @@ public class InternalTranslogManager implements TranslogManager {
             if (tracker != null) {
                 tracker.markSeqNoAsPersisted(seqNo);
             }
-        }, translogUUID, translogFactory, startedPrimarySupplier);
+        }, translogUUID, translogFactory, startedPrimarySupplier, translogOperationHelper);
         assert translog.getGeneration() != null;
         this.translog = translog;
         assert pendingTranslogRecovery.get() == false : "translog recovery can't be pending before we set it";
@@ -368,7 +369,8 @@ public class InternalTranslogManager implements TranslogManager {
         LongConsumer persistedSequenceNumberConsumer,
         String translogUUID,
         TranslogFactory translogFactory,
-        BooleanSupplier startedPrimarySupplier
+        BooleanSupplier startedPrimarySupplier,
+        TranslogOperationHelper translogOperationHelper
     ) throws IOException {
         return translogFactory.newTranslog(
             translogConfig,
@@ -377,7 +379,8 @@ public class InternalTranslogManager implements TranslogManager {
             globalCheckpointSupplier,
             primaryTermSupplier,
             persistedSequenceNumberConsumer,
-            startedPrimarySupplier
+            startedPrimarySupplier,
+            translogOperationHelper
         );
     }